requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-tools==1.70.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.2.2
pandas==2.2.3
protobuf==5.29.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
SQLAlchemy==2.0.37
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2025.1
Werkzeug==3.1.3
